Fix ChoiceEnum to correctly subclass the Enum class
Initial implementation of `ChoiceEnum` subclassed `object` instead of
`Enum` itself since Python was throwing errors. The errors was mostly
because ChoiceEnum was declaring a new class member `CUSTOM_NAMES` which
isn't allowed by the Enum class.

This fixes the implementation and modes the custom label logic to a
method that is meant to be overriden by child classes instead