Move game ownership property from Game model to Player
The relationship created in the `Game` was redundant and unnecessary
since it might imply that a `Player` object that belongs to a different
game could be the owner of the game if their ID was used in the `owner`
field of the `Game` model. This definitely does not make sense and so we
make sure that only players of a game could become owners by making it a
propery of the `Player` model instead.