Add `action` method to the base resident model
Although ideally `Role`s should have the `action` method, each action
needs a context (i.e. it operates on a specific game instead of other
roles in general). The `Role` models do not have this information and it
is the `Resident` model that already has references to the current game.

Instead of forcing every call to `action` to pass in a `Game` model (if
it lived on the `Role` model), we'll just put it in the model that has
the context that it needs.